Add tutorial link
There was some people that thought this tool was meant only to Linux users so I hope that with this tutorial they know that everyone can use it. Maybe they were talking about svg2ass - which is mostly meant for GNU/Linux users - and didn't saw the "GUI" part when commenting, but let's assume people are not that stupid and were thinking this tool requires using Linux because the "SVG" in the name.